% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{app_deps}
\alias{app_deps}
\title{The shiny app dependencies}
\usage{
app_deps(path = ".", recursive = TRUE)
}
\arguments{
\item{path}{path to the shiny app. Default: \code{"."}}

\item{recursive}{logical if to assess the dependencies recursively. Default: TRUE}
}
\value{
character vector with dependency packages or data.frame when checking recursively.
}
\description{
the shiny app dependencies packages are checked recursively.
The \code{c("Depends", "Imports", "LinkingTo")} DESCRIPTION files fields are check recursively.
The required dependencies have to be installed in the local repository.
}
\note{
the base packages are not taken into account.
}
\examples{
\dontrun{
# Please update the path to the custom shiny app
app_path <- system.file("examples/04_mpg", package = "shiny")
pacs::app_deps(app_path)
pacs::app_deps(app_path, recursive = FALSE)
}
}
