% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{available_packages}
\alias{available_packages}
\title{List Available Packages at CRAN-like Repositories}
\usage{
available_packages(repos)
}
\arguments{
\item{repos}{character vector, the base URL(s) of the repositories to use. Default \code{pacs::biocran_repos()}}
}
\description{
available_packages returns a matrix of details corresponding to packages currently available at one or more repositories. The current list of packages is downloaded over the internet (or copied from a local mirror).
}
\keyword{internal}
