% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{bio_releases}
\alias{bio_releases}
\title{Retrieving all Bioconductor releases}
\usage{
bio_releases()
}
\value{
data.frame with the same structure as the html table on \verb{https://www.bioconductor.org/about/release-announcements/}.
}
\description{
Retrieving all Bioconductor releases.
The data is downloaded from \verb{https://www.bioconductor.org/about/release-announcements/}.
}
\note{
Results are cached for 30 minutes with \code{memoise} package.
}
\examples{
\dontrun{
pacs::bio_releases()
}
}
