% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{biocran_repos}
\alias{biocran_repos}
\title{Simple wrapper around \code{BiocManager::repositories}}
\usage{
biocran_repos(...)
}
\arguments{
\item{...}{optional \code{BiocManager::repositories} arguments.}
}
\value{
named character vector of repositories.
}
\description{
Simple wrapper around \code{BiocManager::repositories}, suppress messages which are expected e.g. for RStudio users.
}
\examples{
biocran_repos()
biocran_repos(version = "3.13")
}
