% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{checked_packages}
\alias{checked_packages}
\title{Retrieving all R CRAN packages check pages statuses.}
\usage{
checked_packages()
}
\value{
data.frame with the same structure as the html table on \verb{https://cran.r-project.org/web/checks/check_summary_by_package.html}.
}
\description{
Retrieving all R CRAN packages check pages statuses.
The data is downloaded from \verb{https://cran.r-project.org/web/checks/check_summary_by_package.html}.
}
\note{
Results are cached for 1 hour with \code{memoise} package.
Some packages could be duplicated as not all tests are performed for a new version so two versions still coexists.
Checks with asterisks (*) indicate that checking was not fully performed, this is a case for less than 1\% of all packages.
}
\examples{
\dontrun{
checked_packages()
}
}
