% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{checkpage_packages}
\alias{checkpage_packages}
\title{Downloading all R CRAN packages check page statuses.}
\usage{
checkpage_packages()
}
\value{
data.frame
}
\description{
retrieving all R CRAN packages check pages.
The data is downloaded from \verb{https://cran.r-project.org/web/checks/check_summary_by_package.html}.
}
\note{
Results are cached for 1 hour with \code{memoise} package.
}
