% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_identification.R
\name{compare_models}
\alias{compare_models}
\title{Use gLRT to compare two models and obtain p values}
\usage{
compare_models(
  x_full,
  theta_estimated_full,
  x_null,
  theta_estimated_null,
  q_vec,
  c_by_r,
  df_test,
  mc.cores = 1
)
}
\arguments{
\item{x_full}{Design matrix for the full model}

\item{theta_estimated_full}{Estimated coefficients for the full model}

\item{x_null}{Design matrix for the null model}

\item{theta_estimated_null}{Estimated coefficients for the null model}

\item{q_vec}{A vector of capturing probability for each cell}

\item{c_by_r}{Cell by region matrix}

\item{df_test}{Degree of freedom of the test}

\item{mc.cores}{Number of cores in multi-core computing}
}
\value{
A vector of p values based on gLRT
}
\description{
Use gLRT to compare two models and obtain p values
}
