% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compareVersionsMax}
\alias{compareVersionsMax}
\title{Maximum version across vector}
\usage{
compareVersionsMax(vec)
}
\arguments{
\item{vec}{character vector}
}
\value{
character
}
\description{
Reduce function over the utils::compareVersion
}
