% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compareVersionsMin}
\alias{compareVersionsMin}
\title{Minimum version across vector}
\usage{
compareVersionsMin(vec)
}
\arguments{
\item{vec}{character vector}
}
\value{
character
}
\description{
Reduce function over the utils::compareVersion
}
