% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param-estimate_logit_get_p_by_t_June.R
\name{compute_infor_mat_tilda}
\alias{compute_infor_mat_tilda}
\title{Title compute_information matrix tilda}
\usage{
compute_infor_mat_tilda(xdumm, p_bg, q_vec, y_vec)
}
\arguments{
\item{xdumm}{A design matrix, X}

\item{p_bg}{open probability vector}

\item{q_vec}{A vector of capturing probability}

\item{y_vec}{A vector of observed open (1) or close (0) state for each cell,
must match the length of q}
}
\value{
information matrix tilda
}
\description{
Title compute_information matrix tilda
}
