% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{cran_flavors}
\alias{cran_flavors}
\title{Retrieving all R CRAN servers flavors}
\usage{
cran_flavors()
}
\value{
data.frame with the same structure as the html table on \verb{https://cran.r-project.org/web/checks/check_flavors.html}.
}
\description{
Retrieving all R CRAN servers flavors.
The data is downloaded from \verb{https://cran.r-project.org/web/checks/check_flavors.html}.
}
\note{
Results are cached for 1 hour with \code{memoise} package.
}
\examples{
\dontrun{
cran_flavors()
}
}
