% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dir_size}
\alias{dir_size}
\title{Size of the package}
\usage{
dir_size(path = ".", recursive = TRUE)
}
\arguments{
\item{path}{\code{character} path to the shiny app. Default: \code{"."}}

\item{recursive}{\code{logical} if to assess the dependencies recursively. Default: TRUE}
}
\value{
\code{numeric} size in bytes, to get MB ten divide by \code{10**6}.
}
\description{
size of package.
}
\examples{
\dontrun{
cat(pacs::dir_size(system.file(package = "stats")) / 10**6, "MB")
}
}
