% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param-estimate_logit_get_p_by_t_June.R
\name{estimate_parameters}
\alias{estimate_parameters}
\title{Estimate parameters for the full model}
\usage{
estimate_parameters(
  r_by_c,
  design_mat,
  par_initial,
  cap_rate_vec,
  mc.cores = 1
)
}
\arguments{
\item{r_by_c}{Region by cell matrix}

\item{design_mat}{Desing matrix}

\item{par_initial}{Initialized parameters}

\item{cap_rate_vec}{A vector of capturing probability}

\item{mc.cores}{Number of multi-cores, default = 1}
}
\value{
Estimated coefficient for the full model
}
\description{
Estimate parameters for the full model
}
