% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param-estimate_logit_get_p_by_t_June.R
\name{estimate_parameters_null}
\alias{estimate_parameters_null}
\title{Estimate parameters for the null model}
\usage{
estimate_parameters_null(
  r_by_c,
  design_mat,
  par_initial,
  hold_zero,
  cap_rate_vec,
  mc.cores = 1
)
}
\arguments{
\item{r_by_c}{Region by cell matrix}

\item{design_mat}{Desing matrix}

\item{par_initial}{Initialized parameters}

\item{hold_zero}{Which parameter to hold zero during estimation}

\item{cap_rate_vec}{A vector of capturing probability}

\item{mc.cores}{Number of multi-cores, default = 1}
}
\value{
Estimated coefficient for the null model
}
\description{
Estimate parameters for the null model
}
