% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PACS_test_Seurat.R
\name{FindMarkersPACS}
\alias{FindMarkersPACS}
\title{Differential test with PACS on Seurat object}
\usage{
FindMarkersPACS(
  object,
  ident.1 = NULL,
  ident.2 = NULL,
  latent.vars = NULL,
  group.by = NULL,
  subset.ident = NULL,
  assay = "count",
  reduction = NULL,
  ...
)
}
\arguments{
\item{object}{A \code{Seurat} object}

\item{ident.1}{Identity class to define markers for; pass an object of class
\code{phylo} or 'clustertree' to find markers for a node in a cluster tree;
passing 'clustertree' requires \code{BuildClusterTree} to have been run}

\item{ident.2}{A second identity class for comparison; if \code{NULL},
use all other cells for comparison; if an object of class \code{phylo} or
'clustertree' is passed to \code{ident.1}, must pass a node to find markers for}

\item{latent.vars}{Latent variables to be controlled for from the meta.data
column of the Seurat object}

\item{group.by}{Regroup cells into a different identity class prior to
performing differential expression (see example)}

\item{subset.ident}{Subset a particular identity class prior to regrouping.
Only relevant if group.by is set (see example)}

\item{assay}{Assay to use in differential expression testing, for PACS, we
use \code{count} as assay}

\item{reduction}{Reduction to use in differential expression testing - will
test for DE on cell embeddings}

\item{...}{Additional arguments for identifying differential peaks}
}
\description{
Run differential test using the similar style as in Seurat
\code{Seurat::FindMarkers()} function.
Note 1: This function is modified from Seurat package v5.1.0
Note 2: Seurat must be installed to use this function.
}
\concept{differential_expression}
