% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param-estimate_logit_get_p_by_t_June.R
\name{infor_mat}
\alias{infor_mat}
\title{Title Information matrix -- original without Firth prior}
\usage{
infor_mat(xdumm, p_bg, q_vec)
}
\arguments{
\item{xdumm}{A design matrix, X with dummy variables}

\item{p_bg}{A vector of open probability}

\item{q_vec}{A vector of capturing rates}
}
\value{
information matrix
}
\description{
Title Information matrix -- original without Firth prior
}
