% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param-estimate_logit_get_p_by_t_June.R
\name{irls_iter_null}
\alias{irls_iter_null}
\title{Iteratively reweighted least squares function (main function) for the
null model}
\usage{
irls_iter_null(
  y_vec,
  xdumm,
  theta_estimated,
  hold_zero,
  stop_criteria = 1e-06,
  q_vec
)
}
\arguments{
\item{y_vec}{Observed counts}

\item{xdumm}{Design matrix}

\item{theta_estimated}{Estimated coefficient values}

\item{hold_zero}{Which element to hold zero in the null model}

\item{stop_criteria}{Resolution to stop, default = 1e-6}

\item{q_vec}{Capturing probability vector}
}
\value{
Converged coefficient estimate for the null model
}
\description{
Iteratively reweighted least squares function (main function) for the
null model
}
