% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid.R
\name{lib_validate}
\alias{lib_validate}
\title{Compare current and expected packages under .libPaths.}
\usage{
lib_validate(
  lib.loc = NULL,
  fields = c("Depends", "Imports", "LinkingTo"),
  lifeduration = FALSE
)
}
\arguments{
\item{lib.loc}{character. Default: NULL}

\item{fields}{character vector with possible values c("Depends", "Imports", "LinkingTo", "Suggests"). Default: c("Depends", "Imports", "LinkingTo")}

\item{lifeduration}{logical if to add life duration column. Default: FALSE}
}
\value{
data.frame with 5/6 columns Package Version.expected.min Version.have. "" means newest version.
}
\description{
Checking the healthy of the libarary.
}
\note{
Version.expected.min column not count packages which are not a dependency for any package, so could not be find in DESCRIPTION files.
}
\examples{
lib_validate()
}
