% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_identification.R
\name{loss_firth_from_wii}
\alias{loss_firth_from_wii}
\title{Title Compute loss function with Firth regularization from Wii matrix}
\usage{
loss_firth_from_wii(wii_sqrt, xdumm)
}
\arguments{
\item{wii_sqrt}{The square root of Wii diagonal elements as a vector}

\item{xdumm}{Design matrix}
}
\value{
Regularized log loss value
}
\description{
Title Compute loss function with Firth regularization from Wii matrix
}
