% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param-estimate_logit_get_p_by_t_June.R
\name{loss_fun}
\alias{loss_fun}
\title{Title Calculate Loss Function without Firth prior}
\usage{
loss_fun(p_bg, q_vec, y_vec)
}
\arguments{
\item{p_bg}{A vector of open probability for cell type group g}

\item{q_vec}{A vector of capturing rates}

\item{y_vec}{A vector of observed open (1) or close (0) state for each cell,
must match the length of q}
}
\value{
Log loss value without Firth prior
}
\description{
Title Calculate Loss Function without Firth prior
}
