% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param-estimate_logit_get_p_by_t_June.R
\name{loss_fun_star}
\alias{loss_fun_star}
\title{Title Regularized log loss function}
\usage{
loss_fun_star(xdumm, p_bg, q_vec, y_vec)
}
\arguments{
\item{xdumm}{Observed counts}

\item{p_bg}{Vector of open probability}

\item{q_vec}{Capturing probability vector}

\item{y_vec}{Observed counts}
}
\value{
Regularized log loss
}
\description{
Title Regularized log loss function
}
