% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param-estimate_logit_get_p_by_t_June.R
\name{loss_grad_pen}
\alias{loss_grad_pen}
\title{Title Gradient of penalized loss ( Invariant prior)}
\usage{
loss_grad_pen(xdumm, p_bg, q_vec, inf_mat, wii_sqrt_X)
}
\arguments{
\item{xdumm}{A design matrix, X}

\item{p_bg}{A value of open probability}

\item{q_vec}{A vector of capturing probability}

\item{inf_mat}{Information matrix}

\item{wii_sqrt_X}{sqrt(wii) as a vector times the design matrix}
}
\value{
Gradient of penalized loss ( Invariant prior)
}
\description{
Title Gradient of penalized loss ( Invariant prior)
}
