% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{match_flavors}
\alias{match_flavors}
\title{Get all matched CRAN servers to the local \code{OS}}
\usage{
match_flavors()
}
\value{
\code{character} vector matched server names.
}
\description{
CRAN servers matched to the local \code{OS}.
}
\note{
The Internet connection is needed to use the function.
}
\examples{
\dontrun{
pacs::match_flavors()
}
}
