% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_identification.R
\name{our_method_no_firth}
\alias{our_method_no_firth}
\title{Title Internal evaluation use only -- PACS Without Firth}
\usage{
our_method_no_firth(data_matrix_pos, data_matrix_neg, true_q_pos, true_q_neg)
}
\arguments{
\item{data_matrix_pos}{Count matrix in foreground group}

\item{data_matrix_neg}{Count matrix in background group}

\item{true_q_pos}{Capturing probability vector for the foreground group}

\item{true_q_neg}{Capturing probability vector for the background group}
}
\value{
P values for each cell
}
\description{
Title Internal evaluation use only -- PACS Without Firth
}
