% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{pac_bioreleases}
\alias{pac_bioreleases}
\title{Retrieving all Bioconductor releases}
\usage{
pac_bioreleases()
}
\value{
data.frame with the same structure as the html table on \verb{https://www.bioconductor.org/about/release-announcements/}.
}
\description{
Retrieving all Bioconductor releases.
The data is downloaded from \verb{https://www.bioconductor.org/about/release-announcements/}.
}
\note{
Results are cached for 1 hour with \code{memoise} package.
}
\examples{
pac_bioreleases()
}
