% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{pac_bioversion}
\alias{pac_bioversion}
\title{Retrieving most recent Bioconductor releases for current R version}
\usage{
pac_bioversion()
}
\value{
character Bioconductor release.
}
\description{
Retrieving most recent Bioconductor releases for current R version.
}
\examples{
pac_bioversion()
}
