% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{pac_checkpage}
\alias{pac_checkpage}
\title{Retrieving the package R CRAN check page}
\usage{
pac_checkpage(pac)
}
\arguments{
\item{pac}{character a package name.}
}
\value{
data.frame.
}
\description{
Retrieving the R CRAN package check page.
}
\note{
Results are cached for 1 hour with \code{memoise} package.
If you need to check many packages at once then is recommended usage of \code{pacs::checked_packages}.
Please as a courtesy to the R CRAN, don't overload their server by constantly using this function.
}
\examples{
pac_checkpage("dplyr")
}
