% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pac_checkred}
\alias{pac_checkred}
\title{Checking the R CRAN package check page status}
\usage{
pac_checkred(
  pac,
  scope = c("ERROR", "FAIL"),
  repos = "https://cran.rstudio.com/"
)
}
\arguments{
\item{pac}{character a package name.}

\item{scope}{character vector scope of the check, accepted values c("ERROR", "FAIL", "WARN", "NOTE"). Default c("ERROR", "FAIL")}

\item{repos}{character the base URL of the repositories to use. Default \verb{https://cran.rstudio.com/}}
}
\value{
logical if the package fail under specified criteria.
}
\description{
using R CRAN package check page to validate if there are ANY errors and/or warnings and/or notes.
}
\note{
Results are cached for 1 hour with \code{memoise} package.
If you need to check many packages at once then recommended usage of \code{pacs::checkpage_packages}.
}
\examples{
pac_checkred("dplyr")
pac_checkred("dplyr", scope = c("ERROR"))
}
