% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comapre.R
\name{pac_compare_exports}
\alias{pac_compare_exports}
\title{Compare NAMESPACE exports between specific packages versions}
\usage{
pac_compare_exports(
  pac,
  old = NULL,
  new = NULL,
  lib.loc = NULL,
  repos = "https://cran.rstudio.com/"
)
}
\arguments{
\item{pac}{character a package name.}

\item{old}{character an old version of package.}

\item{new}{character a new version of package.}

\item{lib.loc}{character. Default: NULL}

\item{repos}{character the base URL of the repositories to use. Default \verb{https://cran.rstudio.com/}}
}
\value{
list with two slots, added and removed
}
\description{
using the remote github CRAN mirror to compare NAMESPACE exports between specific packages versions.
}
\note{
The comparison is only about exports.
}
\examples{
pac_compare_exports("memoise", "0.2.1", "2.0.0")
pac_compare_exports("shiny", "1.4.0", "1.6.0")
}
