% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{pac_deps}
\alias{pac_deps}
\title{Package dependencies from DESCRIPTIONS files.}
\usage{
pac_deps(
  pac,
  fields = c("Depends", "Imports", "LinkingTo"),
  lib.loc = NULL,
  attr = TRUE
)
}
\arguments{
\item{pac}{character a package name.}

\item{fields}{character vector with possible values c("Depends", "Imports", "LinkingTo", "Suggests"). Default: c("Depends", "Imports", "LinkingTo")}

\item{lib.loc}{character vector. Is omitted for non NULL version., Default: NULL}

\item{attr}{logical specify if pac and its version should be added as a attribute of data.frame or for FALSE as a additional record. Default: TRUE}
}
\value{
data.frame
}
\description{
Package dependencies from DESCRIPTIONS files.
}
\examples{
pac_deps("stats")
}
