% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_remote.R
\name{pac_deps_timemachine}
\alias{pac_deps_timemachine}
\title{Package dependencies for certain version or time point.}
\usage{
pac_deps_timemachine(
  pac,
  version = NULL,
  at = NULL,
  fields = c("Depends", "Imports", "LinkingTo")
)
}
\arguments{
\item{pac}{character a package name.}

\item{version}{character version of package. Default: NULL}

\item{at}{Date old version of package. Default: NULL}

\item{fields}{character vector with possible values c("Depends", "Imports", "LinkingTo", "Suggests"). Default: c("Depends", "Imports", "LinkingTo")}
}
\value{
named vector package dependencies and their versions at the release date + 1/at arg  of main package.
}
\description{
Package dependencies from DESCRIPTIONS files retrieved recursively for certain version or time point.
}
\note{
Longer lived version is taken if 2 is available at the same date (switch time).
}
\examples{
pacs::pac_deps_timemachine("memoise", "0.2.1")
pacs::pac_deps_timemachine("memoise", at = as.Date("2021-01-01"))
}
