% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/description.R
\name{pac_description}
\alias{pac_description}
\title{package DESCRIPTION file}
\usage{
pac_description(
  pac,
  version = NULL,
  at = NULL,
  local = FALSE,
  lib.loc = .libPaths(),
  repos = "https://cran.rstudio.com/"
)
}
\arguments{
\item{pac}{\code{character} a package name.}

\item{version}{\code{character} version of a package. Default: \code{NULL}}

\item{at}{\code{Date} from which to take the version. Default: \code{NULL}}

\item{local}{\code{logical} if to use local repository (or newest remote packages). Default: \code{FALSE}}

\item{lib.loc}{\code{character} vector of search paths with local packages. Default: \code{.libPaths()}}

\item{repos}{\code{character} vector repositories URLs to use. Used only for the validation. Default \verb{https://cran.rstudio.com/}}
}
\value{
\code{list} with names proper for DESCRIPTION file fields.
}
\description{
CRAN package DESCRIPTION file taken locally or remotely from GITHUB CRAN mirror or CRAN website.
By default works for the newest package version.
}
\note{
Results are cached for 30 minutes with \code{memoise} package.
}
\examples{
\dontrun{
pacs::pac_description("dplyr", version = "0.8.0")
pacs::pac_description("dplyr", at = as.Date("2019-02-01"))
}
}
