% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/howold.R
\name{pac_healthy}
\alias{pac_healthy}
\title{Package healty state at a specific Date or for specific version}
\usage{
pac_healthy(pac, version = NULL, at = NULL)
}
\arguments{
\item{pac}{character a package name.}

\item{version}{character version of package. Default: NULL}

\item{at}{Date old version of package. Default: NULL}
}
\value{
logical if package is healthy, for newest release for current state.
If the newest release is published less than 7 days ago then the class of object is "not-sure".
}
\description{
using cran website to get a package version/versions used at a specific Date interval.
A healthy package is a if it was published for more than 7 days.
CRAN team gives around one week to resolved a package which gave errors under the check page.
}
\note{
Function will scrap two CRAN URLS. Works only with CRAN packages.
Please as a courtesy to the R CRAN, don't overload their server by constantly using this function.
To minimize the overload the memoise package was used.
}
\examples{
\dontrun{
pac_healthy("dplyr")
pac_healthy("dplyr", version = "0.8.0")
}
}
