% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pac_last}
\alias{pac_last}
\title{Getting the most recent package version}
\usage{
pac_last(pac, repos = "https://cran.rstudio.com/")
}
\arguments{
\item{pac}{character a package name.}

\item{repos}{character the base URL of the repository to use. Default \verb{https://cran.rstudio.com/}}
}
\value{
character most recent package version.
}
\description{
using \code{utils::available.packages} to get the newest package version.
}
\note{
Results are cached for 1 hour with \code{memoise} package.
}
\examples{
pac_last("dplyr")
}
