% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifeduration.R
\name{pac_lifeduration}
\alias{pac_lifeduration}
\title{Package life duration at specific Date or for a specific version}
\usage{
pac_lifeduration(pac, version = NULL, at = NULL)
}
\arguments{
\item{pac}{character a package name.}

\item{version}{character version of package. Default: NULL}

\item{at}{Date old version of package. Default: NULL}
}
\value{
logical if package is healthy.
}
\description{
using CRAN website to get a package life duration for certain version or at a specific Date.
}
\note{
Function will scrap two CRAN URLS. Works only with CRAN packages.
Please as a courtesy to the R CRAN, don't overload their server by constantly using this function.
Results are cached for 1 hour with \code{memoise} package.
}
\examples{
pac_lifeduration("memoise")
pac_lifeduration("dplyr", version = "0.8.0")
}
