% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespace.R
\name{pac_namespace}
\alias{pac_namespace}
\title{Package NAMESPACE file}
\usage{
pac_namespace(
  pac,
  version = NULL,
  at = NULL,
  local = FALSE,
  lib.loc = NULL,
  repos = "https://cran.rstudio.com/"
)
}
\arguments{
\item{pac}{character a package name.}

\item{version}{character package version. Default: NULL}

\item{at}{Date. Default: NULL}

\item{local}{logical if to use local library. Default: FALSE}

\item{lib.loc}{character used optionally when local is equal TRUE. Default: NULL}

\item{repos}{character the base URL of the repository to use. Used only for the validation. Default \verb{https://cran.rstudio.com/}}
}
\value{
list with names proper for NAMESPACE file, the same as format as returned by \code{base::parseNamespaceFile}.
}
\description{
package NAMESPACE file taken locally or remotely from GITHUB CRAN mirror or CRAN website.
}
\note{
Results are cached for 1 hour with \code{memoise} package.
This function is mainly built under source code from \code{base::parseNamespaceFile}.
}
\examples{
pac_namespace("dplyr", version = "0.8.0")
pac_namespace("dplyr", at = as.Date("2019-02-01"))
pac_namespace("memoise", local = TRUE)

}
