% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/news.R
\name{pac_news}
\alias{pac_news}
\title{Get NEWS for a package}
\usage{
pac_news(
  pac,
  version = NULL,
  at = NULL,
  local = FALSE,
  lib.loc = .libPaths(),
  repos = "https://cran.rstudio.com/"
)
}
\arguments{
\item{pac}{\code{character} a package name.}

\item{version}{\code{character} version of a package. Default: \code{NULL}}

\item{at}{\code{Date} from which to take the version. Default: \code{NULL}}

\item{local}{\code{logical} if to use local repository (or newest remote packages). Default: \code{FALSE}}

\item{lib.loc}{\code{character} vector of search paths with local packages. Default: \code{.libPaths()}}

\item{repos}{\code{character} vector repositories URLs to use. Used only for the validation. Default \verb{https://cran.rstudio.com/}}
}
\value{
\code{character} with NEWS content.
}
\description{
Get NEWS for a package from CRAN or local
}
\note{
Results are cached for 30 minutes with \code{memoise} package.
}
\examples{
\dontrun{
pacs::pac_news("dplyr", version = "0.8.0")
pacs::pac_news("dplyr", at = as.Date("2019-02-01"))
}
}
