% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onlast.R
\name{pac_on}
\alias{pac_on}
\title{Checking if a packge is in repositories}
\usage{
pac_on(pac, repos = biocran_repos())
}
\arguments{
\item{pac}{character a package name.}

\item{repos}{character vector base URLs of the repositories to use. By default checking CRAN and Bioconductor. Default \code{pacs::biocran_repos()}}
}
\value{
logical if a package is inside ropositories.
}
\description{
using \code{utils::available.packages} to check if package is in repoitories.
}
\note{
Results are cached for 1 hour with \code{memoise} package.
}
\examples{
pac_on("dplyr")
pac_on("dplyr", "https://cran.rstudio.com/" )
pac_on("dplyr", biocran_repos()[grep("Bio", names(biocran_repos()))])
}
