% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timemachine.R
\name{pac_timemachine}
\alias{pac_timemachine}
\title{Package metadata for all releases}
\usage{
pac_timemachine(
  pac,
  at = NULL,
  from = NULL,
  to = NULL,
  version = NULL,
  source = c("crandb", "cran")
)
}
\arguments{
\item{pac}{character a package name.}

\item{at}{Date old version of package. Default: NULL}

\item{from}{Date new version of package. Default: NULL}

\item{to}{Date CRAN URL. Default: NULL}

\item{version}{character version of package. Default: NULL}

\item{source}{character one of \code{c("crandb", "cran")}. Using the \code{MEATCRAN} DB or the direct web page download from CRAN. Default: \code{"crandb"}}
}
\value{
data.frame with 7 columns
\describe{
\item{Package}{character package name.}
\item{Version}{character package version.}
\item{Released}{character release Date}
\item{Archived}{character archived Date.}
\item{LifeDuration}{\code{difftime} number of days the version was the newest one.}
\item{URL}{character the suffix of the base URL to \code{tar.gz} file.}
\item{Size}{character size of the \code{tar.gz} file.}
}
}
\description{
Using CRAN website to get a package metadata used at a specific Date or a Date interval or for specific version.
}
\note{
Function will scrap two CRAN URLS. Works only with CRAN packages.
Please as a courtesy to the R CRAN, don't overload their servers by constantly using this function.
The base part of URL in the result is \verb{https://cran.r-project.org/src/contrib/}.
Results are cached for 30 minutes with \code{memoise} package.
The \code{crandb} R packages database is a part of \code{METACRAN} project, source:
Csárdi G, Salmon M (2022). \code{pkgsearch}: Search and Query CRAN R Packages. \verb{https://github.com/r-hub/pkgsearch}, \verb{https://r-hub.github.io/pkgsearch/}.
}
\examples{
\dontrun{
pac_timemachine("dplyr", at = as.Date("2017-02-02"))
pac_timemachine("dplyr", from = as.Date("2017-02-02"), to = as.Date("2018-04-02"))
pac_timemachine("dplyr", at = Sys.Date())
pac_timemachine("tidyr", from = as.Date("2020-06-01"), to = Sys.Date())
}
}
