% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/howold.R
\name{pac_versions}
\alias{pac_versions}
\title{Package version at a specific Date or a Date interval}
\usage{
pac_versions(pac, at = NULL, from = NULL, to = NULL)
}
\arguments{
\item{pac}{character a package name.}

\item{at}{Date old version of package. Default: NULL}

\item{from}{Date new version of package. Default: NULL}

\item{to}{Date cran URL. Default: NULL}
}
\value{
data.frame
}
\description{
using cran website to get a package version/versions used at a specific Date or a Date interval.
}
\note{
Function will scrap two CRAN URLS. Works only with CRAN packages.
}
\examples{
\dontrun{
pac_versions("dplyr", at = as.Date("2017-02-02"))
pac_versions("dplyr", from = as.Date("2017-02-02"), to = as.Date("2018-04-02"))
pac_versions("dplyr", at = Sys.Date())
}
}
