% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pacs_base}
\alias{pacs_base}
\title{Get base R packages}
\usage{
pacs_base(startup = FALSE)
}
\arguments{
\item{startup}{\code{logical} include only \code{startup} packages. Default: FALSE}
}
\value{
\code{character} vector
}
\description{
get base packages, all or only \code{startup}.
}
\examples{
\dontrun{
pacs_base()
pacs_base(startup = TRUE)
}
}
