% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pacs_checkred}
\alias{pacs_checkred}
\title{Checking the R CRAN packages check pages statuses}
\usage{
pacs_checkred(
  pacs,
  scope = c("ERROR", "WARN", "NOTE"),
  repos = "https://cran.rstudio.com/"
)
}
\arguments{
\item{pacs}{character vector packages names.}

\item{scope}{character vector scope of the check, accepted values c("ERROR", "WARN", "NOTE"). Default c("ERROR", "WARN")}

\item{repos}{character the base URL of the repositories to use. Default \verb{https://cran.rstudio.com/}}
}
\value{
logical named vector if packages fail under specified criteria.
}
\description{
using R CRAN packages check pages to validate if there are ANY error and/or warning and/or note.
}
\note{
Results are cached for 1 hour with \code{memoise} package.
}
\examples{
pacs_checkred(c("dplyr", "devtools"))
pacs_checkred("dplyr", scope = c("ERROR"))
}
