% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{pacs_deps}
\alias{pacs_deps}
\title{compare specific package versions}
\usage{
pacs_deps(pacs = NULL, lib.loc = NULL)
}
\arguments{
\item{pacs}{character vector}

\item{lib.loc}{character, Default: NULL}
}
\value{
data.frame
}
\description{
using remote
}
\examples{

}
