% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/description.R
\name{pacs_description}
\alias{pacs_description}
\title{packages DESCRIPTION files}
\usage{
pacs_description(
  pacs,
  versions = NULL,
  at = NULL,
  local = FALSE,
  lib.loc = NULL
)
}
\arguments{
\item{pacs}{character vector packages names.}

\item{versions}{character vector versions. Default: NULL}

\item{at}{Date . Default: NULL}

\item{local}{logical if to use local library. Default: FALSE}

\item{lib.loc}{character used optionally when local is equal TRUE. Default: NULL}
}
\value{
list of list with names proper for DESCRIPTION file fields.
}
\description{
packages DESCRIPTION files taken locally or if remotely then from GITHUB CRAN mirror or CRAN website.
}
\examples{
pacs_description(c("dplyr", "memoise"), version = c("0.8.1", "1.0.0"))
pacs_description(c("dplyr", "memoise"), at = as.Date("2019-02-01"))
}
