% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/howold.R
\name{pacs_health}
\alias{pacs_health}
\title{Packages health state at a specific Date or for a specific versions}
\usage{
pacs_health(pacs, versions = NULL, at = NULL)
}
\arguments{
\item{pacs}{character vector packages names.}

\item{versions}{character vector versions of packages. Default: NULL}

\item{at}{Date old version of package. Default: NULL}
}
\value{
logical vector, TRUE if a package is healthy.
If the newest release is published less than 7 days ago then the class of object is "not-sure" for newest version.
}
\description{
using cran website to get a package version/versions used at a specific Date interval.
A healthy package is a if it was published for more than 7 days.
CRAN team gives around one week to resolved a package which gave errors under the check page.
}
\note{
Function will scrap two CRAN URLS. Works only with CRAN packages.
Please as a courtesy to the R CRAN, don't overload their server by constantly using this function.
To minimize the overload the memoise package was used, to cache the data.
}
\examples{
\dontrun{
pacs_health(c("dplyr", "devtools"))
pacs_health(c("dplyr", "devtools"), versions = c("0.8.0", "2.4.0"))
}
}
