% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifeduration.R
\name{pacs_lifedurations}
\alias{pacs_lifedurations}
\title{Packages life duration for a specific version}
\usage{
pacs_lifedurations(
  pacs,
  versions,
  source = c("crandb", "loop_crandb", "loop_cran"),
  lib.loc = .libPaths(),
  repos = biocran_repos()
)
}
\arguments{
\item{pacs}{character vector packages names.}
}
\description{
using CRAN website to get a package life duration for certain version or at a specific Date.
}
