% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timemachine.R
\name{pacs_timemachine}
\alias{pacs_timemachine}
\title{Packages versions at a specific Date or a Date interval}
\usage{
pacs_timemachine(pacs, at = NULL, from = NULL, to = NULL)
}
\arguments{
\item{pacs}{character vector packages names.}

\item{at}{Date old version of package. Default: NULL}

\item{from}{Date new version of package. Default: NULL}

\item{to}{Date CRAN URL. Default: NULL}
}
\value{
data.frame
}
\description{
using CRAN website to get packages version/versions used at a specific Date or a Date interval.
}
\note{
Function will scrap two CRAN URLS. Works only with CRAN packages.
For bigger lists might need a few minutes.
Please as a courtesy to the R CRAN, don't overload their server by constantly using this function.
The base part of URL in the result is \verb{https://cran.r-project.org/src/contrib/}.
}
\examples{
pacs_timemachine(c("dplyr", "memoise"), from = as.Date("2018-06-30"), to = as.Date("2019-01-01"))
pacs_timemachine(c("dplyr", "memoise"), at = Sys.Date())

}
