% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{pacs_validate}
\alias{pacs_validate}
\title{Compare current and expected packages under .libPaths.}
\usage{
pacs_validate(
  pacs,
  lib.loc = NULL,
  fields = c("Depends", "Imports", "LinkingTo")
)
}
\arguments{
\item{pacs}{character vector packages names.}

\item{lib.loc}{character. Default: NULL}

\item{fields}{character vector with possible values c("Depends", "Imports", "LinkingTo", "Suggests"). Default: c("Depends", "Imports", "LinkingTo")}
}
\value{
data.frame with 3 columns Package Version.expected.min Version.have. "" means newest version.
}
\description{
Checking the healthy of the libarary.
}
\note{
Version.expected.min column not count packages which are not a dependency for any package, so could not be find in DESCRIPTION files.
}
\examples{
\dontrun{
pacs_validate(c("devtools", "renv"))
}
}
