% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/howold.R
\name{pacs_versions}
\alias{pacs_versions}
\title{Packages versions at a specific Date or a Date interval}
\usage{
pacs_versions(pacs, at = NULL, from = NULL, to = NULL)
}
\arguments{
\item{pacs}{character vector packages names.}

\item{at}{Date old version of package. Default: NULL}

\item{from}{Date new version of package. Default: NULL}

\item{to}{Date cran URL. Default: NULL}
}
\value{
data.frame
}
\description{
using cran website to get packages version/versions used at a specific Date or a Date interval.
}
\note{
Function will scrap two CRAN URLS. Works only with CRAN packages. For bigger lists might need a few minutes.
}
\examples{
\dontrun{
pacs_versions(c("dplyr", "shiny"), from = as.Date("2018-06-30"), to = as.Date("2019-01-01"))
pacs_versions(c("dplyr", "shiny"), at = Sys.Date())
}
}
