% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_github_file.R
\name{read_github_file}
\alias{read_github_file}
\title{Read a file from a GitHub CRAN repository}
\usage{
read_github_file(pac, version, file, repos = "https://cran.rstudio.com/")
}
\arguments{
\item{pac}{\code{character} a package name.}

\item{version}{\code{character} version of a package. Default: \code{NULL}}

\item{file}{\code{character} file name to read. Possible values are \code{DESCRIPTION}, \code{NEWS} and \code{NAMESPACE}.}

\item{repos}{\code{character} vector of repositories URLs to use. By default checking CRAN and newest Bioconductor per R version. Default \code{pacs::biocran_repos()}}
}
\description{
Read a file from a GitHub CRAN repository.
}
\note{
if the file is not found in the GitHub repository, it will try to find it in the CRAN archive.
}
\keyword{internal}
