% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size.R
\name{shiny_app_size}
\alias{shiny_app_size}
\title{Size of the shiny app}
\usage{
shiny_app_size(path = ".")
}
\arguments{
\item{path}{path to the shiny app. Default: \code{"."}}
}
\value{
numeric size in bytes, to get MB ten divide by \code{10**6}.
}
\description{
size of the shiny app dependencies packages and the app directory.
The app dependencies packages are checked recursively.
}
\examples{
\dontrun{
# Please update the path to the shiny app
cat(pacs::shiny_app_size(system.file("examples/04_mpg", package = "shiny")) / 10**6, "MB")
}
}
