% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size.R
\name{shiny_true_size}
\alias{shiny_true_size}
\title{Shiny app true size}
\usage{
shiny_true_size(path = getwd())
}
\arguments{
\item{path}{character path to shiny app. Default: getwd()}
}
\value{
numeric sizes in bytes.
}
\description{
shiny app true size as it takes into account all its dependencies.
}
