% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_validate.R
\name{validate_compare_input}
\alias{validate_compare_input}
\title{validate compare input}
\usage{
validate_compare_input(pac, old, new, lib.loc, repos)
}
\description{
validate compare input
}
\keyword{internal}
