% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{validate_lib}
\alias{validate_lib}
\title{Compare current and expected packages under .libPaths.}
\usage{
validate_lib(lib.loc = NULL, fields = c("Depends", "Imports", "LinkingTo"))
}
\arguments{
\item{lib.loc}{character. Default: NULL}

\item{fields}{character vector with possible values c("Depends", "Imports", "LinkingTo", "Suggests"). Default: c("Depends", "Imports", "LinkingTo")}
}
\value{
data.frame with 3 columns Package Version.expected.min Version.have. "" means newest version.
}
\description{
Checking the healthy of the libarary.
}
\note{
Version.expected.min column not count packages which are not a dependency for any package, so could not be find in DESCRIPTION files.
}
\examples{
validate_lib()
}
