% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid.R
\name{validate_online}
\alias{validate_online}
\title{Append a data.frame with online attributes}
\usage{
validate_online(
  result,
  version_name_new = "Version.expected",
  lifeduration,
  checkred,
  repos,
  lib.loc = .libPaths()
)
}
\description{
internal function to append a data.frame with an online related sources.
The input data.frame has a specific structure, such contain \code{c("Package", version_name_new)} columns.
}
\keyword{internal}
