% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{app_deps}
\alias{app_deps}
\title{The shiny app dependencies}
\usage{
app_deps(
  path = ".",
  fields = c("Depends", "Imports", "LinkingTo"),
  lib.loc = .libPaths(),
  local = TRUE,
  base = FALSE,
  description_v = FALSE,
  recursive = TRUE,
  repos = biocran_repos()
)
}
\arguments{
\item{path}{\code{character} path to the shiny app. Default: \code{"."}}

\item{fields}{\code{character} vector listing the types of dependencies, a subset of \code{c("Depends", "Imports", "LinkingTo", "Suggests", "Enhances")}.
Character string "all" is shorthand for that vector, character string "most" for the same vector without "Enhances", character string "strong" (default) for the first three elements of that vector.
Default: \code{c("Depends", "Imports", "LinkingTo")}}

\item{lib.loc}{\code{character} vector of search paths with local packages. Default: \code{.libPaths()}}

\item{local}{\code{logical} if to use local repository (or newest remote packages). Default: TRUE}

\item{base}{\code{logical} if to add base packages too. If \code{TRUE} then \code{pacs::pacs_base()} are taken into account. Default: \code{FALSE}}

\item{description_v}{\code{logical} if the dependencies version should be taken from description files, minimal required. By default installed versions are taken. Default: \code{FALSE}}

\item{recursive}{\code{logical} if to assess the dependencies recursively. Default: TRUE}

\item{repos}{\code{character} vector base URLs of the repositories to use. By default checking CRAN and newest Bioconductor per R version. Default \code{pacs::biocran_repos()}}
}
\value{
\code{character} vector with dependency packages or data.frame when checking recursively.
}
\description{
the shiny app dependencies packages are checked recursively.
The \code{c("Depends", "Imports", "LinkingTo")} DESCRIPTION files fields are checked recursively.
The required dependencies have to be installed in the local repository.
The default arguments setup is recommended.
}
\note{
\code{renv} package has to be installed.
}
\examples{
\dontrun{
library(renv)
# Please update the path to the custom shiny app
app_path <- system.file("examples/04_mpg", package = "shiny")
pacs::app_deps(app_path)
pacs::app_deps(app_path, recursive = FALSE)
}
}
