% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size.R
\name{app_size}
\alias{app_size}
\title{Size of the shiny app}
\usage{
app_size(path = ".")
}
\arguments{
\item{path}{path to the shiny app. Default: \code{"."}}
}
\value{
numeric size in bytes, to get MB ten divide by \code{10**6}.
}
\description{
The size of shiny app is a sum of dependencies packages and the app directory.
The app dependencies packages are checked recursively.
}
\examples{
\dontrun{
# Please update the path to the shiny app
cat(pacs::app_size(system.file("examples/04_mpg", package = "shiny")) / 10**6, "MB")
}
}
