% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{biocran_repos}
\alias{biocran_repos}
\title{CRAN and Bioconductor repositories}
\usage{
biocran_repos(version = pac_bioversion())
}
\arguments{
\item{version}{character the Bioconductor release.
By default the newest Bioconductor release for the specific R version is assumed. The maximum version is equal to \code{head(pacs::pac_bioreleases()$Release, 1)}. Default \code{pacs::pac_bioversion()}}
}
\value{
named character vector of repositories.
}
\description{
CRAN and Bioconductor repositories.
The newest Bioconductor release for the specific R version is assumed.
}
\examples{
biocran_repos()
biocran_repos(version = "3.13")
}
