% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{biocran_repos}
\alias{biocran_repos}
\title{CRAN and Bioconductor repositories}
\usage{
biocran_repos(bioversion = NULL)
}
\arguments{
\item{bioversion}{character the Bioconductor release.
By default the newest Bioconductor release for the specific R version is assumed, if not available only CRAN repository is returned.
The maximum version could be checked with \code{pacs::bio_releases()}. Default NULL}
}
\value{
named character vector of repositories.
}
\description{
CRAN and Bioconductor repositories.
The newest Bioconductor release for the specific R version is assumed.
}
\examples{
biocran_repos()
}
