% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compareVersionsMin}
\alias{compareVersionsMin}
\title{Minimum version across the vector}
\usage{
compareVersionsMin(vec, na.rm = TRUE)
}
\arguments{
\item{vec}{character vector}

\item{na.rm}{logical if to remove NA values.}
}
\value{
character minimal version
}
\description{
Reduce function over the \code{utils::compareVersion}
}
\examples{
compareVersionsMin(c("1.1.1", "0.2.0"))
}
