% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_type_annotation.R
\name{estimate_label_no_cap_rate_all_pks}
\alias{estimate_label_no_cap_rate_all_pks}
\title{Estimate cell type label for new data using all peaks as input}
\usage{
estimate_label_no_cap_rate_all_pks(r_by_t, in_r_by_c, pks_sel, alpha = 1)
}
\arguments{
\item{r_by_t}{Region-by cell type matrix generated from a annotated dataset}

\item{in_r_by_c}{Input (unannotated) region by cell matrix}

\item{pks_sel}{selected peaks as informative for cell type label prediction}

\item{alpha}{Weight for negative peaks, default = 1}
}
\value{
A matrix of cell types by cell matrix, with elements representing
probability of being in that cell type
}
\description{
We will identify relevant peaks
}
