% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param-estimate_logit_get_p_by_t_June.R
\name{infor_mat}
\alias{infor_mat}
\title{Title Information matrix -- original without Firth prior}
\usage{
infor_mat(xdumm, p_bg, q_vec)
}
\arguments{
\item{xdumm}{A design matrix, X with dummy variables}

\item{q_vec}{A vector of capturing rates}

\item{theta_estimated}{Estimated values of \theta}

\item{y_vec}{A vector of observed open (1) or close (0) state for each cell,
must match the length of q}
}
\value{
information matrix
}
\description{
Title Information matrix -- original without Firth prior
}
