% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param-estimate_logit_get_p_by_t_June.R
\name{irls_iter}
\alias{irls_iter}
\title{Title Iteratively reweighted least squares function (main function)}
\usage{
irls_iter(y_vec, xdumm, theta_estimated, stop_criteria = 1e-06, q_vec)
}
\arguments{
\item{y_vec}{Observed counts}

\item{xdumm}{Design matrix}

\item{theta_estimated}{Estimated coefficient values}

\item{stop_criteria}{Resolution to stop, default = 1e-6}

\item{q_vec}{Capturing probability vector}
}
\value{
Converged coefficient estimate
}
\description{
Title Iteratively reweighted least squares function (main function)
}
