% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pac_checkred}
\alias{pac_checkred}
\title{Checking the R CRAN package check page status}
\usage{
pac_checkred(pac, scope = c("ERROR", "WARN"))
}
\arguments{
\item{pac}{character a package name.}

\item{scope}{character vector scope of the check, accepted values c("ERROR", "WARN", "NOTE"). Default c("ERROR", "WARN")}
}
\value{
logical if the package fail under specified criteria.
}
\description{
using R CRAN package check page to validate if there are ANY error and/or warning and/or note.
}
\examples{
pac_checkred("dplyr")
pac_checkred("dplyr", scope = c("ERROR"))
}
