% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{pac_compare_versions}
\alias{pac_compare_versions}
\title{compare dependencies of specific package versions}
\usage{
pac_compare_versions(pac, old, new, repos = "http://cran.rstudio.com/")
}
\arguments{
\item{pac}{character a package name.}

\item{old}{character old version of package.}

\item{new}{character new version of package.}

\item{repos}{character cran URL. Default: "http://cran.rstudio.com/"}
}
\value{
data.frame
}
\description{
using remote cran to compare specific package versions dependencies.
}
\note{
Function will temporarily download two packages.
}
\examples{
\dontrun{
pac_compare_versions("shiny", "1.4.0-2", "1.6.0")
}
}
