% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{pac_compare_versions}
\alias{pac_compare_versions}
\title{compare specific package versions}
\usage{
pac_compare_versions(pac, old, new, repos = "http://cran.rstudio.com/")
}
\arguments{
\item{pac}{character}

\item{old}{character}

\item{new}{character}

\item{repos}{character}
}
\value{
data.frame
}
\description{
using remote
}
\note{
will temporarily download two packages
}
\examples{

}
