% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comapre.R
\name{pac_compare_versions}
\alias{pac_compare_versions}
\title{Compare dependencies of specific packages versions}
\usage{
pac_compare_versions(
  pac,
  old,
  new,
  fields = c("Imports", "Depends", "LinkingTo")
)
}
\arguments{
\item{pac}{character a package name.}

\item{old}{character an old version of package.}

\item{new}{character a new version of package.}

\item{fields}{character a vector with possible values \code{c("Depends", "Imports", "LinkingTo", "Suggests")}. Default: \code{c("Depends", "Imports", "LinkingTo")}}
}
\value{
data.frame with 4 columns.
}
\description{
using remote CRAN to compare specific packages versions dependencies.
}
\examples{
pac_compare_versions("memoise", "0.2.1", "2.0.0")
}
