% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{pac_deps}
\alias{pac_deps}
\title{Package dependencies from DESCRIPTIONS files.}
\usage{
pac_deps(
  pac,
  fields = c("Depends", "Imports", "LinkingTo"),
  lib.loc = NULL,
  base = FALSE,
  local = TRUE,
  description_v = FALSE,
  attr = TRUE
)
}
\arguments{
\item{pac}{character a package name.}

\item{fields}{character vector with possible values c("Depends", "Imports", "LinkingTo", "Suggests"). Default: c("Depends", "Imports", "LinkingTo")}

\item{lib.loc}{character vector. Is omitted for non NULL version., Default: NULL}

\item{base}{logical if to add base packages too. Default: FALSE}

\item{local}{logical if to use newest CRAN packages, where by default local ones are used. Default: TRUE}

\item{description_v}{if the dependecies version should be taken from description files, minimal required. Default: FALSE}

\item{attr}{logical specify if pac and its version should be added as a attribute of data.frame or for FALSE as a additional record. Default: TRUE}
}
\value{
data.frame with packages and their versions. Versions are taken from \code{installed.packages} or newest released.
}
\description{
Package dependencies from DESCRIPTIONS files.
}
\examples{
pacs::pac_deps("stats", base = TRUE)$Package
\dontrun{
pacs::pac_deps("shiny")$Package
pacs::pac_deps("shiny", description_v = FALSE)
}
}
