% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{pac_deps}
\alias{pac_deps}
\title{specific package dependencies from DESCRIPTIONS files.}
\usage{
pac_deps(
  pac,
  fields = c("Depends", "Imports", "LinkingTo"),
  lib.loc = NULL,
  version = NULL,
  attr = TRUE,
  repos = "http://cran.rstudio.com/"
)
}
\arguments{
\item{pac}{character}

\item{fields}{character vector, Default: c("Depends", "Imports", "LinkingTo")}

\item{lib.loc}{character vector, Default: NULL}

\item{version}{character, Default: NULL}

\item{attr}{logical , Default: NULL}

\item{repos}{character, Default: "http://cran.rstudio.com/"}
}
\value{
data.frame
}
\description{
xxx
}
\examples{

}
