% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{pac_deps_dev}
\alias{pac_deps_dev}
\title{Package dependencies - developer perspective}
\usage{
pac_deps_dev(pac)
}
\arguments{
\item{pac}{character a package name.}
}
\description{
A higher-level function, build from \code{pacs::pacs_deps}.
Package dependencies installed when checking the package.
\verb{"Depends", "Imports", "LinkingTo", "Suggests"} fields from the DESCRIPTION file and
their recursive dependencies taken from \verb{"Depends", "Imports", "LinkingTo"} fields.
Taken remotely for the newest version.
}
\examples{
\dontrun{
pacs::pac_deps_dev("dplyr")
pacs::pac_deps_dev("cat2cat")
}
}
