% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{pac_deps_dev}
\alias{pac_deps_dev}
\title{package dependencies - developer perspective}
\usage{
pac_deps_dev(pac)
}
\arguments{
\item{pac}{character a package name.}
}
\description{
Package dependencies installed when checking the package.
}
\examples{
\dontrun{
  pac_deps_dev("dplyr")
  pac_deps_dev("cat2cat")
}
}
