% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{pac_deps_dev}
\alias{pac_deps_dev}
\title{Package dependencies - developer perspective}
\usage{
pac_deps_dev(pac, base = FALSE, attr = TRUE, repos = pacs::biocran_repos())
}
\arguments{
\item{pac}{character a package name.}

\item{base}{logical if to add base packages too.}

\item{attr}{logical specify if a package and its version should be added as a attribute of data.frame or for FALSE as a additional record. Default: TRUE}

\item{repos}{character vector URLs of the repositories to use. By default checking CRAN and newest Bioconductor per R version. Default \code{pacs::biocran_repos()}
If \code{TRUE} then \code{pacs::pacs_base()} are taken into account. Default: FALSE}
}
\description{
A higher-level function, build from \code{pacs::pacs_deps}.
Package dependencies installed when checking the package.
\verb{"Depends", "Imports", "LinkingTo", "Suggests"} fields from the DESCRIPTION file and
their recursive dependencies taken from \verb{"Depends", "Imports", "LinkingTo"} fields.
Dependencies are taken remotely for the newest version.
}
\examples{
\dontrun{
pacs::pac_deps_dev("dplyr")
pacs::pac_deps_dev("cat2cat")
}
}
