% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{pac_deps_heavy}
\alias{pac_deps_heavy}
\title{Package direct dependencies and number of dependencies for each of them}
\usage{
pac_deps_heavy(
  pac,
  lib.loc = .libPaths(),
  base = FALSE,
  local = FALSE,
  repos = pacs::biocran_repos()
)
}
\arguments{
\item{pac}{character a package name.}

\item{lib.loc}{character vector, used optionally when local is equal TRUE. Default: \code{.libPaths()}}

\item{base}{logical if to add base packages too. If \code{TRUE} then \code{pacs::pacs_base()} are taken into account. Default: FALSE}

\item{local}{logical if to use local repository or newest CRAN packages, where by default local packages are used. Default: TRUE}

\item{repos}{character vector URLs of the repositories to use. By default checking CRAN and newest Bioconductor per R version. Default \code{pacs::biocran_repos()}}
}
\value{
\verb{named integer} number of dependencies for each of the direct one.
}
\description{
A higher-level function, build from \code{pacs::pacs_deps} and \code{tools::package_dependencies}.
A tool to identify a main sources of dependencies.
}
\examples{
\dontrun{
pacs::pac_deps_heavy("dplyr")
pacs::pac_deps_heavy("pacs")
}
}
