% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{pac_deps_timemachine}
\alias{pac_deps_timemachine}
\title{Package dependencies for certain version or time point.}
\usage{
pac_deps_timemachine(
  pac,
  version = NULL,
  at = NULL,
  fields = c("Depends", "Imports", "LinkingTo")
)
}
\arguments{
\item{pac}{character a package name.}

\item{version}{character version of package. Default: NULL}

\item{at}{Date old version of package. Default: NULL}

\item{fields}{character vector with possible values c("Depends", "Imports", "LinkingTo", "Suggests"). Default: c("Depends", "Imports", "LinkingTo")}
}
\value{
named vector package dependencies and their versions at the release date + 1 of main package.
}
\description{
Package dependencies from DESCRIPTIONS files retrieved recursively for certain version or time point.
}
\note{
Longer lived version is taken if 2 is available at the same date (switch time).
}
\examples{
pacs::pac_deps_timemachine("miceFast", "0.6.2")
pacs::pac_deps_timemachine("miceFast", at = as.Date("2019-01-01"))
}
