% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{pac_deps_user}
\alias{pac_deps_user}
\title{package dependencies - user perspective}
\usage{
pac_deps_user(pac)
}
\arguments{
\item{pac}{character a package name.}
}
\description{
Package dependencies installed when run \code{installed.packages}.
}
\examples{
\dontrun{
  pac_deps_user("dplyr")
  pac_deps_user("cat2cat")
}
}
