% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{pac_deps_user}
\alias{pac_deps_user}
\title{Package dependencies - user perspective}
\usage{
pac_deps_user(pac)
}
\arguments{
\item{pac}{character a package name.}
}
\description{
A higher-level function, build from \code{pacs::pacs_deps}.
Package dependencies installed when run \code{installed.packages}.
\verb{"Depends", "Imports", "LinkingTo"} fields from the DESCRIPTION file and
their recursive dependencies taken from the same fields.
Taken remotely for the newest version.
}
\examples{
\dontrun{
pacs::pac_deps_user("dplyr")
pacs::pac_deps_user("cat2cat")
}
}
