% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{pac_deps_user}
\alias{pac_deps_user}
\title{Package dependencies - user perspective}
\usage{
pac_deps_user(
  pac,
  base = FALSE,
  local = FALSE,
  attr = TRUE,
  repos = pacs::biocran_repos()
)
}
\arguments{
\item{pac}{\code{character} a package name.}

\item{base}{\code{logical} if to add base packages too. If \code{TRUE} then \code{pacs::pacs_base()} are taken into account. Default: FALSE}

\item{local}{logical if to use local repository (or newest remote packages). Default: FALSE}

\item{attr}{\code{logical} specify if a package and its version should be added as an attribute of data.frame or for FALSE as an additional record. Default: TRUE}

\item{repos}{\code{character} vector base URLs of the repositories to use. By default checking CRAN and newest Bioconductor per R version. Default \code{pacs::biocran_repos()}}
}
\value{
\code{data.frame} with packages and their versions. Versions are taken from \code{installed.packages} or newest released.
}
\description{
A higher-level function, build from \code{pacs::pacs_deps}.
Package dependencies installed when run \code{installed.packages}.
\verb{"Depends", "Imports", "LinkingTo"} fields from the DESCRIPTION file and
their recursive dependencies taken from the same fields.
Dependencies are taken remotely for the newest version.
}
\examples{
\dontrun{
pacs::pac_deps_user("dplyr")
pacs::pac_deps_user("pacs")
# with the main package in the list
pacs::pac_deps_user("pacs", attr = FALSE)
}
}
