% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/description.R
\name{pac_description}
\alias{pac_description}
\title{Package DESCRIPTION file}
\usage{
pac_description(
  pac,
  version = NULL,
  at = NULL,
  local = FALSE,
  lib.loc = NULL,
  repos = "https://cran.rstudio.com/"
)
}
\arguments{
\item{pac}{character a package name.}

\item{version}{character package version. Default: NULL}

\item{at}{Date. Default: NULL}

\item{local}{logical if to use local library. Default: FALSE}

\item{lib.loc}{character used optionally when local is equal TRUE. Default: NULL}

\item{repos}{character the base URL of the repositories to use. Default \verb{https://cran.rstudio.com/}}
}
\value{
list with names proper for DESCRIPTION file fields.
}
\description{
package DESCRIPTION file taken locally or remotely from GITHUB CRAN mirror or CRAN website.
}
\note{
Results are cached for 1 hour with \code{memoise} package.
}
\examples{
pac_description("dplyr", version = "0.8.0")
pac_description("dplyr", at = as.Date("2019-02-01"))
}
