% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.R
\name{pac_description}
\alias{pac_description}
\title{package DESCRIPTION file}
\usage{
pac_description(pac, version = NULL, at = NULL, local = FALSE, lib.loc = NULL)
}
\arguments{
\item{pac}{character a package name.}

\item{version}{character package version. Default: NULL}

\item{at}{Date. Default: NULL}

\item{local}{logical if to use local library. Default: FALSE}

\item{lib.loc}{character used optionally when local is equal TRUE. Default: NULL}
}
\value{
numeric size in bytes.
}
\description{
package DESCRIPTION file taken from CRAN website.
}
\examples{
# Will fail
# pac_description("dplyr", version = "0.8.0")
# pac_description("dplyr", at = as.Date("2019-02-01"))
}
