% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifeduration.R
\name{pac_health}
\alias{pac_health}
\title{Package health state at a specific Date or for a specific version}
\usage{
pac_health(pac, version = NULL, at = NULL, limit = 7)
}
\arguments{
\item{pac}{character a package name.}

\item{version}{character version of package. Default: NULL}

\item{at}{Date old version of package. Default: NULL}

\item{limit}{numeric at least days to treat as healthy. Default: 7}
}
\value{
logical if package is healthy.
If the newest release is published less than 7 days ago then the class of object is "not-sure".
}
\description{
using CRAN website to get a package version/versions used at a specific Date interval.
A healthy package is a if it was published for more than 7 days.
CRAN team gives around one week to resolved a package which gave errors under the check page.
}
\note{
Function will scrap two CRAN URLS. Works only with CRAN packages.
Please as a courtesy to the R CRAN, don't overload their server by constantly using this function.
}
\examples{
pac_health("dplyr")
pac_health("dplyr", version = "0.8.0")

}
