% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifeduration.R
\name{pac_lifeduration}
\alias{pac_lifeduration}
\title{Package life duration at specific Date or for a specific version}
\usage{
pac_lifeduration(pac, version = NULL, at = NULL)
}
\arguments{
\item{pac}{character a package name.}

\item{version}{character version of package. Default: NULL}

\item{at}{Date old version of package. Default: NULL}
}
\value{
logical if package is healthy.
If the newest release is published less than 7 days ago then the class of object is "not-sure".
}
\description{
using CRAN website to get a package version/versions used at a specific Date interval.
}
\note{
Function will scrap two CRAN URLS. Works only with CRAN packages.
Please as a courtesy to the R CRAN, don't overload their server by constantly using this function.
}
\examples{
pac_lifeduration("memoise")
pac_lifeduration("dplyr", version = "0.8.0")

}
