% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespace.R
\name{pac_namespace}
\alias{pac_namespace}
\title{package NAMESPACE file}
\usage{
pac_namespace(
  pac,
  version = NULL,
  at = NULL,
  local = FALSE,
  lib.loc = .libPaths(),
  repos = "https://cran.rstudio.com/"
)
}
\arguments{
\item{pac}{character a package name.}

\item{version}{character package version, By default the is_installed version. Default: NULL}

\item{at}{Date. Default: NULL}

\item{local}{logical if to use local library. Default: FALSE}

\item{lib.loc}{character vector, used optionally when local is equal TRUE. Default: \code{.libPaths()}}

\item{repos}{character the base URL of the CRAN repository to use. Used only for the validation. Default \verb{https://cran.rstudio.com/}}
}
\value{
list with names proper for NAMESPACE file, the same as format as returned by \code{base::parseNamespaceFile}.
}
\description{
CRAN package NAMESPACE file taken locally or remotely from GITHUB CRAN mirror or CRAN website.
}
\note{
Results are cached for 30 minutes with \code{memoise} package.
This function is mainly built under source code from \code{base::parseNamespaceFile}.
}
\examples{
\dontrun{
pacs::pac_namespace("dplyr", version = "0.8.0")
pacs::pac_namespace("dplyr", at = as.Date("2019-02-01"))
pacs::pac_namespace("memoise", local = TRUE)
}
}
