% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size.R
\name{pac_size}
\alias{pac_size}
\title{Size of the package}
\usage{
pac_size(pac, lib.loc = .libPaths())
}
\arguments{
\item{pac}{\code{character} a package name.}

\item{lib.loc}{\code{character} vector of search paths for local packages. Default: \code{.libPaths()}}
}
\value{
\code{numeric} size in bytes, to get MB ten divide by \code{10**6}.
}
\description{
Size of package.
}
\examples{
\dontrun{
cat(pacs::pac_size("stats") / 10**6, "MB")
}
}
