% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size.R
\name{pac_size}
\alias{pac_size}
\title{Size of the package}
\usage{
pac_size(pac, lib.loc = NULL)
}
\arguments{
\item{pac}{character a package name.}

\item{lib.loc}{character vector. Default: NULL}
}
\value{
numeric size in bytes, to get MB ten divide by \code{10**6}.
}
\description{
size of package.
}
\examples{
cat(pacs::pac_size("stats") / 10**6, "MB")
}
