% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timemachine.R
\name{pac_timemachine}
\alias{pac_timemachine}
\title{Package version at a specific Date or a Date interval}
\usage{
pac_timemachine(
  pac,
  at = NULL,
  from = NULL,
  to = NULL,
  version = NULL,
  repos = "https://cran.rstudio.com/"
)
}
\arguments{
\item{pac}{character a package name.}

\item{at}{Date old version of package. Default: NULL}

\item{from}{Date new version of package. Default: NULL}

\item{to}{Date CRAN URL. Default: NULL}

\item{version}{character version of package. Default: NULL}

\item{repos}{character the base URL of the repository to use. Used only for the validation. Default \verb{https://cran.rstudio.com/}}
}
\value{
data.frame with 7 columns
\describe{
\item{Package}{character package name.}
\item{Version}{character package version.}
\item{Released}{character release Date}
\item{Archived}{character archived Date.}
\item{LifeDuration}{\code{difftime} number of days the version was the newest one.}
\item{URL}{character the suffix of the base URL to tar.gz file.}
\item{Size}{character size of the \code{tar.gz} file.}
}
}
\description{
using CRAN website to get a package version/versions used at a specific Date or a Date interval.
}
\note{
Function will scrap two CRAN URLS. Works only with CRAN packages.
Please as a courtesy to the R CRAN, don't overload their servers by constantly using this function.
The base part of URL in the result is \verb{https://cran.r-project.org/src/contrib/}.
Results are cached for 1 hour with \code{memoise} package.
}
\examples{
pac_timemachine("dplyr", at = as.Date("2017-02-02"))
pac_timemachine("dplyr", from = as.Date("2017-02-02"), to = as.Date("2018-04-02"))
pac_timemachine("dplyr", at = Sys.Date())
}
