% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{pac_true_size}
\alias{pac_true_size}
\title{Package true size}
\usage{
pac_true_size(
  pac,
  version = NULL,
  fields = c("Depends", "Imports", "LinkingTo"),
  lib.loc = NULL
)
}
\arguments{
\item{pac}{character a package name.}

\item{version}{character version of the package. Other than NULL triggers remote mode and omit lib.loc. Default: NULL}

\item{fields}{character vector, Default: c("Depends", "Imports", "LinkingTo")}

\item{lib.loc}{character vector, Default: NULL}
}
\value{
data.frame
}
\description{
Package true size as it takes into account dependencies, optionally for remote cran version.
}
\examples{
# size in Mb
pacs::pac_true_size("stats")/10**6
}
