% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{pac_true_size}
\alias{pac_true_size}
\title{Package true size}
\usage{
pac_true_size(
  pac,
  fields = c("Depends", "Imports", "LinkingTo"),
  lib.loc = NULL,
  base = FALSE
)
}
\arguments{
\item{pac}{character a package name.}

\item{fields}{character vector, Default: c("Depends", "Imports", "LinkingTo")}

\item{lib.loc}{character vector, Default: NULL}

\item{base}{logical if to add base packages too. Default: FALSE}
}
\value{
data.frame
}
\description{
Package true size as it takes into account dependencies.
}
\examples{
# size in Mb
pacs::pac_true_size("stats")/10**6
}
