% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size.R
\name{pac_true_size}
\alias{pac_true_size}
\title{Package true size}
\usage{
pac_true_size(
  pac,
  fields = c("Depends", "Imports", "LinkingTo"),
  lib.loc = NULL,
  base = FALSE,
  exclude_joint = 0L
)
}
\arguments{
\item{pac}{character a package name.}

\item{fields}{character vector, Default: c("Depends", "Imports", "LinkingTo")}

\item{lib.loc}{character vector, Default: NULL}

\item{base}{logical if to add base packages too. Default: FALSE}

\item{exclude_joint}{integer exclude packages which are dependencies of at least N other packages. Default: 0}
}
\value{
numeric sizes in bytes.
}
\description{
Package true size as it takes into account dependencies.
}
\examples{
# size in Mb
pacs::pac_true_size("stats") / 10**6
# exclude packages if at least one other package use it too
\dontrun{
pacs::pac_true_size("devtools", exclude_joint = 1L)/10**6
}
}
