% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{pac_validate}
\alias{pac_validate}
\title{Compare current and expected packages under .libPaths.}
\usage{
pac_validate(
  pac,
  lib.loc = NULL,
  fields = c("Depends", "Imports", "LinkingTo")
)
}
\arguments{
\item{pac}{character a package name.}

\item{lib.loc}{character. Default: NULL}

\item{fields}{character vector with possible values c("Depends", "Imports", "LinkingTo", "Suggests"). Default: c("Depends", "Imports", "LinkingTo")}
}
\value{
data.frame with 3 columns Package Version.expected.min Version.have. "" means newest version.
}
\description{
Checking the healthy of the libarary.
}
\note{
Version.expected.min column not count packages which are not a dependency for any package, so could not be find in DESCRIPTION files.
}
\examples{
\dontrun{
pac_validate("devtools")
}
}
