% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pacs_base}
\alias{pacs_base}
\title{List of base R packages}
\usage{
pacs_base(startup = FALSE)
}
\arguments{
\item{startup}{logical include only startup packages. Default: FALSE}
}
\value{
character vector
}
\description{
using installed.packages and priority equal "base" to retrieve base packages.
}
