% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pacs_checkred}
\alias{pacs_checkred}
\title{Checking the R CRAN packages check pages statuses}
\usage{
pacs_checkred(pacs, scope = c("ERROR", "WARN", "NOTE"))
}
\arguments{
\item{pacs}{character vector packages names.}

\item{scope}{character vector scope of the check, accepted values c("ERROR", "WARN", "NOTE"). Default c("ERROR", "WARN")}
}
\value{
logical named vector if packages fail under specified criteria.
}
\description{
using R CRAN packages check pages to validate if there are ANY error and/or warning and/or note.
}
\examples{
pacs_checkred(c("dplyr", "devtools"))
pacs_checkred("dplyr", scope = c("ERROR"))
}
