% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{pacs_deps}
\alias{pacs_deps}
\title{Packages dependencies}
\usage{
pacs_deps(
  pacs = NULL,
  fields = c("Depends", "Imports", "LinkingTo"),
  lib.loc = NULL,
  attr = TRUE,
  base = FALSE,
  local = TRUE,
  description_v = FALSE
)
}
\arguments{
\item{pacs}{character vector of packages.}

\item{fields}{character vector with possible values \code{c("Depends", "Imports", "LinkingTo", "Suggests")}. Default: \code{c("Depends", "Imports", "LinkingTo")}}

\item{lib.loc}{character vector. Is omitted for non NULL version., Default: NULL}

\item{attr}{logical specify if package and its version should be added as a attribute of data.frame or for FALSE as a additional record. Default: FALSE}

\item{base}{logical if to add base packages too. Default: FALSE}

\item{local}{logical if to use newest CRAN packages, where by default local ones are used. Default: TRUE}

\item{description_v}{if the dependencies version should be taken from description files, minimal required. Default: FALSE}
}
\value{
data.frame
}
\description{
Package dependencies from DESCRIPTION files with installed or expected versions or newest released.
}
\examples{
pacs_deps(c("stats", "base"), base = TRUE, attr = FALSE)
}
