% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.R
\name{pacs_description}
\alias{pacs_description}
\title{packages DESCRIPTION files}
\usage{
pacs_description(
  pacs,
  versions = NULL,
  at = NULL,
  local = FALSE,
  lib.loc = NULL
)
}
\arguments{
\item{pacs}{character vector packages names.}

\item{versions}{character vector versions. Default: NULL}

\item{at}{Date . Default: NULL}

\item{local}{logical if to use local library. Default: FALSE}

\item{lib.loc}{character used optionally when local is equal TRUE. Default: NULL}
}
\value{
numeric size in bytes.
}
\description{
packages DESCRIPTION files taken from CRAN website.
}
\examples{
# pacs_description(c("dplyr", "shiny"), version = c("0.8.1", "1.5.0"))
# pacs_description(c("dplyr", "shiny"), at = as.Date("2019-02-01"))
}
