% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifeduration.R
\name{pacs_lifeduration}
\alias{pacs_lifeduration}
\title{Packages versions life duration at specific Date or for a specific version}
\usage{
pacs_lifeduration(
  pacs,
  versions = NULL,
  at = NULL,
  lib.loc = NULL,
  repos = "https://cran.rstudio.com/"
)
}
\arguments{
\item{pacs}{character vector packages names.}

\item{versions}{character vector versions of packages. Default: NULL}

\item{at}{Date old version of package. Default: NULL}

\item{lib.loc}{character vector. Is omitted for non NULL version. Default: NULL}

\item{repos}{character the base URL of the repositories to use. Default \verb{https://cran.rstudio.com/}}
}
\value{
list named of \code{difftime}, number of days package version was the newest one.
}
\description{
using CRAN website to get packages life duration for versions or specific Date.
}
\note{
Function will scrap two CRAN URLS. Works only with CRAN packages.
Please as a courtesy to the R CRAN, don't overload their server by constantly using this function.
Results are cached for 1 hour with \code{memoise} package, memory cache.
There is used \code{parallel::mclapply} function.
Remember that \code{mclaply} under Windows works like the regular \code{lapply} function.
To set higher number of cores use code like \code{options(mc.cores = parallel::detectCores() - 2)} at the beginning of the session.
}
\examples{
pacs_lifeduration(c("memoise", "rlang"))
pacs_lifeduration("dplyr", version = "0.8.0")
}
