% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifeduration.R
\name{pacs_lifeduration}
\alias{pacs_lifeduration}
\title{Packages life duration for a specific version}
\usage{
pacs_lifeduration(
  pacs,
  versions,
  source = c("crandb", "loop_crandb", "loop_cran"),
  lib.loc = .libPaths(),
  repos = biocran_repos()
)
}
\arguments{
\item{pacs}{character vector packages names.}

\item{versions}{character vector packages versions.}

\item{source}{character one of \code{c("crandb", "loop_crandb", "loop_cran")}. Default: \code{"crandb"}}

\item{lib.loc}{character vector. Is omitted for non NULL version. Default: \code{.libPaths()}}

\item{repos}{character the base CRAN URL of the repository to use. Default \code{"https://cran.rstudio.org"}}
}
\description{
using CRAN website to get a package life duration for certain version or at a specific Date.
}
