% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{pacs_size}
\alias{pacs_size}
\title{size of packages}
\usage{
pacs_size(pacs, lib.loc = NULL)
}
\arguments{
\item{pacs}{character vector}

\item{lib.loc}{character vector, Default: NULL}
}
\value{
data.frame
}
\description{
using remote
}
\examples{


}
