% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size.R
\name{pacs_size}
\alias{pacs_size}
\title{size of packages.}
\usage{
pacs_size(pacs = NULL, lib.loc = NULL)
}
\arguments{
\item{pacs}{character vector packages.}

\item{lib.loc}{character vector. Default: NULL}
}
\value{
named vector of sizes in bytes.
}
\description{
size of packages.
}
\examples{
cat(pacs::pacs_size("stats")/10**6, "Mb")
}
