% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid.R
\name{pacs_validate}
\alias{pacs_validate}
\title{Validate specific packages}
\usage{
pacs_validate(
  pacs,
  lib.loc = NULL,
  fields = c("Depends", "Imports", "LinkingTo")
)
}
\arguments{
\item{pacs}{character vector packages names.}

\item{lib.loc}{character. Default: NULL}

\item{fields}{character vector with possible values \code{c("Depends", "Imports", "LinkingTo", "Suggests")}. Default: \code{c("Depends", "Imports", "LinkingTo")}}
}
\value{
data.frame with 7 columns Package Version.expected.min Version.have. "" means newest version.
}
\description{
Checking the healthy of the specific packages.
}
\note{
Version.expected.min column not count packages which are not a dependency for any package, so could not be find in DESCRIPTION files.
}
\examples{
pacs_validate(c("memoise", "rlang"))
}
