% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{validate_lib}
\alias{validate_lib}
\title{Compare current and expected packages under .libPaths.}
\usage{
validate_lib(lib.loc = NULL)
}
\arguments{
\item{lib.loc}{character. Default: NULL}
}
\value{
data.frame
}
\description{
Checking the healthy of the libarary.
}
\examples{
\dontrun{
validate_lib()
}
}
